+++ src/braket/pennylane_plugin/translation.py
@@ -11,12 +11,14 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
+import numbers
 from functools import reduce, singledispatch
-from typing import Any, FrozenSet, List, Tuple, Union
+from typing import Any, FrozenSet, List, Optional, Tuple, Union
 
 import pennylane as qml
-from braket.circuits import Gate, ResultType, gates, noises, observables
+from braket.circuits import FreeParameter, Gate, ResultType, gates, noises, observables
 from braket.circuits.result_types import (
+    AdjointGradient,
     DensityMatrix,
     Expectation,
     Probability,
@@ -29,6 +31,7 @@ from braket.tasks import GateModelQuantumTaskResult
 from pennylane import numpy as np
 from pennylane.measurements import ObservableReturnTypes
 from pennylane.operation import Observable, Operation
+from pennylane.ops import Adjoint
 
 from braket.pennylane_plugin.ops import (
     MS,
@@ -50,11 +53,11 @@ _BRAKET_TO_PENNYLANE_OPERATIONS = {
     "rx": "RX",
     "rz": "RZ",
     "s": "S",
-    "si": "S.inv",
+    "si": "Adjoint(S)",
     "v": "SX",
-    "vi": "SX.inv",
+    "vi": "Adjoint(SX)",
     "t": "T",
-    "ti": "T.inv",
+    "ti": "Adjoint(T)",
     "cnot": "CNOT",
     "cy": "CY",
     "cz": "CZ",
@@ -109,16 +112,49 @@ def supported_operations(device: Device) -> FrozenSet[str]:
     )
 
 
-def translate_operation(operation: Operation, *args, **kwargs) -> Gate:
+def translate_operation(
+    operation: Operation,
+    use_unique_params: bool = False,
+    param_names: Optional[List[str]] = None,
+    *args,
+    **kwargs,
+) -> Gate:
     """Translates a PennyLane ``Operation`` into the corresponding Braket ``Gate``.
 
     Args:
         operation (Operation): The PennyLane ``Operation`` to translate
+        use_unique_params (bool): If true, numeric parameters in the resulting operation will be
+        replaced with FreeParameter objects (with names corresponding to param_names). Non-numeric
+        parameters will be skipped.
+        param_names (Optional[List[str]]): A list of parameter names
+            to be supplied to the new operation.
 
     Returns:
         Gate: The Braket gate corresponding to the given operation
     """
-    parameters = [p.numpy() if isinstance(p, qml.numpy.tensor) else p for p in operation.parameters]
+    if use_unique_params:
+        param_names = param_names or []
+        parameters = []
+        name_index = 0
+        for param in operation.parameters:
+            # PennyLane passes any non-keyword argument in the operation.parameters list.
+            # In some cases, like the unitary gate or qml.QubitChannel (Kraus noise), these
+            # parameter can be matrices. Braket only supports parameterization of numeric parameters
+            # (so far, these are all angle parameters), so non-numeric parameters are handled
+            # separately.
+            if isinstance(param, numbers.Number):
+                new_param = FreeParameter(param_names[name_index])
+                name_index += 1
+            elif isinstance(param, qml.numpy.tensor):
+                new_param = param.numpy()
+            else:
+                new_param = param
+            parameters.append(new_param)
+    else:
+        parameters = [
+            p.numpy() if isinstance(p, qml.numpy.tensor) else p for p in operation.parameters
+        ]
+
     return _translate_operation(operation, parameters)
 
 
@@ -161,17 +197,17 @@ def _(_: qml.ECR, _parameters):
 
 @_translate_operation.register
 def _(s: qml.S, _parameters):
-    return gates.Si() if s.inverse else gates.S()
+    return gates.S()
 
 
 @_translate_operation.register
 def _(sx: qml.SX, _parameters):
-    return gates.Vi() if sx.inverse else gates.V()
+    return gates.V()
 
 
 @_translate_operation.register
 def _(t: qml.T, _parameters):
-    return gates.Ti() if t.inverse else gates.T()
+    return gates.T()
 
 
 @_translate_operation.register
@@ -207,31 +243,31 @@ def _(_: qml.Toffoli, _parameters):
 @_translate_operation.register
 def _(rx: qml.RX, parameters):
     phi = parameters[0]
-    return gates.Rx(-phi) if rx.inverse else gates.Rx(phi)
+    return gates.Rx(phi)
 
 
 @_translate_operation.register
 def _(ry: qml.RY, parameters):
     phi = parameters[0]
-    return gates.Ry(-phi) if ry.inverse else gates.Ry(phi)
+    return gates.Ry(phi)
 
 
 @_translate_operation.register
 def _(rz: qml.RZ, parameters):
     phi = parameters[0]
-    return gates.Rz(-phi) if rz.inverse else gates.Rz(phi)
+    return gates.Rz(phi)
 
 
 @_translate_operation.register
 def _(phase_shift: qml.PhaseShift, parameters):
     phi = parameters[0]
-    return gates.PhaseShift(-phi) if phase_shift.inverse else gates.PhaseShift(phi)
+    return gates.PhaseShift(phi)
 
 
 @_translate_operation.register
 def _(qubit_unitary: qml.QubitUnitary, parameters):
     U = np.asarray(parameters[0])
-    return gates.Unitary(U.conj().T) if qubit_unitary.inverse else gates.Unitary(U)
+    return gates.Unitary(U)
 
 
 @_translate_operation.register
@@ -280,60 +316,60 @@ def _(_: qml.QubitChannel, parameters):
 @_translate_operation.register
 def _(c_phase_shift: qml.ControlledPhaseShift, parameters):
     phi = parameters[0]
-    return gates.CPhaseShift(-phi) if c_phase_shift.inverse else gates.CPhaseShift(phi)
+    return gates.CPhaseShift(phi)
 
 
 @_translate_operation.register
 def _(c_phase_shift_00: CPhaseShift00, parameters):
     phi = parameters[0]
-    return gates.CPhaseShift00(-phi) if c_phase_shift_00.inverse else gates.CPhaseShift00(phi)
+    return gates.CPhaseShift00(phi)
 
 
 @_translate_operation.register
 def _(c_phase_shift_01: CPhaseShift01, parameters):
     phi = parameters[0]
-    return gates.CPhaseShift01(-phi) if c_phase_shift_01.inverse else gates.CPhaseShift01(phi)
+    return gates.CPhaseShift01(phi)
 
 
 @_translate_operation.register
 def _(c_phase_shift_10: CPhaseShift10, parameters):
     phi = parameters[0]
-    return gates.CPhaseShift10(-phi) if c_phase_shift_10.inverse else gates.CPhaseShift10(phi)
+    return gates.CPhaseShift10(phi)
 
 
 @_translate_operation.register
 def _(iswap: qml.ISWAP, _parameters):
-    return gates.PSwap(3 * np.pi / 2) if iswap.inverse else gates.ISwap()
+    return gates.ISwap()
 
 
 @_translate_operation.register
 def _(pswap: PSWAP, parameters):
     phi = parameters[0]
-    return gates.PSwap(-phi) if pswap.inverse else gates.PSwap(phi)
+    return gates.PSwap(phi)
 
 
 @_translate_operation.register
 def _(xy: qml.IsingXY, parameters):
     phi = parameters[0]
-    return gates.XY(-phi) if xy.inverse else gates.XY(phi)
+    return gates.XY(phi)
 
 
 @_translate_operation.register
 def _(xx: qml.IsingXX, parameters):
     phi = parameters[0]
-    return gates.XX(-phi) if xx.inverse else gates.XX(phi)
+    return gates.XX(phi)
 
 
 @_translate_operation.register
 def _(yy: qml.IsingYY, parameters):
     phi = parameters[0]
-    return gates.YY(-phi) if yy.inverse else gates.YY(phi)
+    return gates.YY(phi)
 
 
 @_translate_operation.register
 def _(zz: qml.IsingZZ, parameters):
     phi = parameters[0]
-    return gates.ZZ(-phi) if zz.inverse else gates.ZZ(phi)
+    return gates.ZZ(phi)
 
 
 @_translate_operation.register
@@ -345,13 +381,42 @@ def _(_gpi: GPi, parameters):
 @_translate_operation.register
 def _(gpi2: GPi2, parameters):
     phi = parameters[0]
-    return gates.GPi2(phi + np.pi) if gpi2.inverse else gates.GPi2(phi)
+    return gates.GPi2(phi)
 
 
 @_translate_operation.register
 def _(ms: MS, parameters):
     phi_0, phi_1 = parameters[:2]
-    return gates.MS(phi_0 + np.pi, phi_1) if ms.inverse else gates.MS(phi_0, phi_1)
+    return gates.MS(phi_0, phi_1)
+
+
+@_translate_operation.register
+def _(adjoint: Adjoint, parameters):
+    if isinstance(adjoint.base, qml.ISWAP):
+        # gates.ISwap.adjoint() returns a different value
+        return gates.PSwap(3 * np.pi / 2)
+    base = _translate_operation(adjoint.base, parameters)
+    if len(base.adjoint()) > 1:
+        raise NotImplementedError(
+            f"The adjoint of the Braket operation {base} contains more than one operation."
+        )
+    return base.adjoint()[0]
+
+
+def get_adjoint_gradient_result_type(
+    observable: Observable,
+    targets: Union[List[int], List[List[int]]],
+    supported_result_types: FrozenSet[str],
+    parameters: List[str],
+):
+    if "AdjointGradient" not in supported_result_types:
+        raise NotImplementedError("Unsupported return type: AdjointGradient")
+    braket_observable = _translate_observable(observable)
+
+    braket_observable = (
+        braket_observable.item() if hasattr(braket_observable, "item") else braket_observable
+    )
+    return AdjointGradient(observable=braket_observable, target=targets, parameters=parameters)
 
 
 def translate_result_type(
@@ -402,7 +467,21 @@ def translate_result_type(
 
 @singledispatch
 def _translate_observable(observable):
-    raise TypeError(f"Unsupported observable: {observable}")
+    raise TypeError(f"Unsupported observable: {type(observable)}")
+
+
+@_translate_observable.register
+def _(H: qml.Hamiltonian):
+    # terms is structured like [C, O] where C is a tuple of all the coefficients, and O is
+    # a tuple of all the corresponding observable terms (X, Y, Z, H, etc or a tensor product
+    # of them)
+    coefficents, pl_observables = H.terms()
+    braket_observables = list(map(lambda obs: _translate_observable(obs), pl_observables))
+    braket_hamiltonian = sum(
+        (coef * obs for coef, obs in zip(coefficents[1:], braket_observables[1:])),
+        coefficents[0] * braket_observables[0],
+    )
+    return braket_hamiltonian
 
 
 @_translate_observable.register
@@ -473,6 +552,20 @@ def translate_result(
     Note:
         Hamiltonian results will be summed over all terms.
     """
+
+    # if braket result contains adjoint gradient, just return it since it should be the only
+    # result type if it's there at all.
+    ag_results = [
+        result for result in braket_result.result_types if result.type.type == "adjoint_gradient"
+    ]
+    if ag_results:
+        ag_result = ag_results[0]
+        key_indices = [int(param_name.split("p_")[1]) for param_name in ag_result.value["gradient"]]
+        return [ag_result.value["expectation"]], [
+            # we need to sort the keys by index since braket can return them in the wrong order
+            ag_result.value["gradient"][f"p_{i}"]
+            for i in sorted(key_indices)
+        ]
     translated = translate_result_type(observable, targets, supported_result_types)
     if isinstance(observable, qml.Hamiltonian):
         coeffs, _ = observable.terms()